refactor: move client version to web3 provider
